requirement.txt
# Core dependencies
torch==2.6.0
torchvision==0.21.0
diffusers==0.33.0.dev0
transformers==4.51.1
accelerate==1.6.0
safetensors==0.5.3

# Gradio for UI
gradio==5.23.3
gradio_client==1.8.0

# PIL and image handling
pillow==11.1.0
numpy==2.2.4

# Hugging Face Hub
huggingface-hub==0.30.2

# Scheduler and runtime utils
filelock==3.18.0
psutil==7.0.0

# Optional but useful
tqdm==4.67.1
requests==2.32.3
packaging==24.2

# Audio for webcam/image inputs (if needed in future)
pydub==0.25.1

# FastAPI backend dependencies (if expanding app)
fastapi==0.115.12
uvicorn==0.34.0
python-multipart==0.0.20
orjson==3.10.16

# Ensure correct CUDA packages
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-curand-cu12==10.3.5.147
nvidia-cufft-cu12==11.2.1.3
nvidia-nccl-cu12==2.21.5
nvidia-nvtx-cu12==12.4.127
nvidia-nvjitlink-cu12==12.4.127

# Optional (image utils)
aiofiles==23.2.1
